### 2023-07-23 #5681 @arfon
@LucVV – thanks for your submission to JOSS. We're managing quite a large backlog of submissions right now so will have to waitlist this while we wait for an editor to become available.
