### 2023-07-23 #5681 @LucVV
Hi @arfon, thank you for informing us about the backlog!

While we wait for an editor, I have already spent some thought on the possible reviewers. To run and test `ZOSPy`, as required in the review process, one will need to have OpticStudio running. Unfortunately, no-one lists that software in their keywords.

We could ask **andibarg** or **Omnistic**, who have `ZOSPy` running. They have however made some contributions to `ZOSPy`, so I'm not sure if they are allowed to review.

Another option would be asking **aquilesC** or **MikeHughesKent** from the list of reviewers. I think they might have access to OpticStudio based on their listed keywords and GitHub profiles.

I hope this helps with finding suitable reviewers!
