### 2023-07-24 9c42b92
Extend test README with a section about reference data generation
