+++ tests/README.md
@@ -39,3 +39,42 @@ These arguments can also be passed to tox, e.g. (note the double dashes `--` aft
 ```shell
 tox run -e py311-standalone -- --output-directory="zospy/is/cool" 
 ```
+
+## Generating test reference data
+
+The unit tests for `zospy.analyses` rely on reference data in order to check the validity of 
+If the unit tests are run for the first time using a specific version of ZOSPy, this data is not yet present.
+It can be generated by running
+
+```shell
+python -m scripts.generate_test_reference_data
+```
+
+The generated reference data files will be added to `tests/data/reference`.
+
+The reference data files are generated using a specification in `scripts/generate_test_reference_data/tests.yaml`.
+If new analys tests are added, this file needs to be extended with a reference specification for these tests.
+A documented example specification is shown below.
+
+```yaml
+- # OpticStudio model as defined in scripts/generate_test_reference_data/systems.py
+  model: simple_system
+  # ZOSPy analysis, specified as a module relative to zospy.analyses
+  analysis: raysandspots.ray_fan
+  # Unit test file, relative to tests/analyses
+  file: test_raysandspots.py
+  # Unit test for which this reference data is intended
+  test: test_ray_fan_returns_correct_result
+  # Parameters that are parametrized using @pytest.mark.parametrize
+  parametrized: [plot_scale, number_of_rays, tangential, sagittal]
+  # List with the parameter sets in parameter_name: parameter_value format
+  parameters:
+    - plot_scale: 0
+      number_of_rays: 20
+      tangential: Aberration_Y
+      sagittal: Aberration_X
+    - plot_scale: 1
+      number_of_rays: 40
+      tangential: Aberration_Y
+      sagittal: Aberration_X
+```
