### 2023-07-24 af1a28a
Update test instructions
