### 2023-07-24 768c6f7
Added function to retrieve inherited object data implementation in nce.py
