### 2023-07-24 efeaa3a
Changes in docstring according to @crnh
