+++ tests/README.md
@@ -42,7 +42,7 @@ tox run -e py311-standalone -- --output-directory="zospy/is/cool"
 
 ## Generating test reference data
 
-The unit tests for `zospy.analyses` rely on reference data in order to check the validity of 
+The unit tests for `zospy.analyses` rely on reference data in order to check the validity of analysis results.
 If the unit tests are run for the first time using a specific version of ZOSPy, this data is not yet present.
 It can be generated by running
 
