### 2023-07-25 270a66f
Updated regex to account for more potential cases
