+++ zospy/analyses/polarization.py
@@ -20,7 +20,8 @@ from zospy.zpcore import OpticStudioSystem
 
 def _get_number_field(name: str, text: str) -> str:
     return re.search(
-        rf"{re.escape(name)}\s*:\s*([-+]?\d+{re.escape(_config.DECIMAL_POINT)}?\d*(?:[Ee][-+]?\d+)?)", text
+        rf"{re.escape(name)}\s*:\s*([-+]?(\d+({re.escape(_config.DECIMAL_POINT)}\d*)?|{re.escape(_config.DECIMAL_POINT)}\d+)(?:[Ee][-+]?\d+)?)",
+        text,
     ).group(1)
 
 
