### 2023-07-25 e89a4ce
added monkeypatch to temporarily change _config.DECIMAL_POINT
Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>
