### 2023-07-25 76b205a
updated names of test functions in TestGetNumberField
