+++ tests/analyses/test_polarization.py
@@ -19,7 +19,7 @@ class TestGetNumberField:
     @pytest.mark.parametrize("exp", _exps)
     @pytest.mark.parametrize("number", _int_numbers)
     @pytest.mark.parametrize("sign", _signs)
-    def test_get_number_field_returns_correct_result_for_integers(self, sign, number, exp):
+    def test_parses_int(self, sign, number, exp):
         number_string = sign + number + exp
         res = _get_number_field("Test", f"Test: {number_string}")
 
@@ -29,9 +29,7 @@ class TestGetNumberField:
     @pytest.mark.parametrize("exp", _exps)
     @pytest.mark.parametrize("number", _float_numbers)
     @pytest.mark.parametrize("sign", _signs)
-    def test_get_number_field_returns_correct_result_for_floats(
-        self, sign, number, exp, decimal_separator, monkeypatch: pytest.MonkeyPatch
-    ):
+    def test_parses_float(self, sign, number, exp, decimal_separator, monkeypatch: pytest.MonkeyPatch):
         monkeypatch.setattr(_config, "DECIMAL_POINT", decimal_separator)
 
         number_string = (sign + number + exp).replace(".", decimal_separator)
