### 2023-07-25 ac3ecaa
changed variable name _exps to _exponents
