### 2023-07-26 ddb78e8
Update conftest.py to test R1.03 for Issue #31
