### 2023-07-26 9937967
Update tests/conftest.py
Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>
