+++ zospy/functions/nce.py
@@ -124,7 +124,7 @@ def get_object_data(obj: _ZOSAPI.Editors.NCE.INCERow) -> _ZOSAPI.Editors.NCE.IOb
     >>> oss = zos.get_primary_system()
     >>> oss.make_nonsequential()
     >>> detector_object = oss.NCE.GetObjectAt(1)
-    >>> detector_type = detector.GetObjectTypeSettings(zp.constants.Editors.NCE.ObjectType.DetectorRectangle)
+    >>> detector_type = detector_object.GetObjectTypeSettings(zp.constants.Editors.NCE.ObjectType.DetectorRectangle)
     >>> detector_object.ChangeType(detector_type)
     >>> detector_data = zp.functions.nce.get_object_data(detector_object)
     >>> number_of_x_pixels = detector_data.NumberXPixels
