### 2023-07-28 dbe7a1b
Fix error in workflow
