+++ .github/workflows/ci.yml
@@ -74,7 +74,7 @@ jobs:
 
   check_file_updates:
     runs-on: ubuntu-latest
-    if: ${{ github.event_name == 'pull_request' }} && contains(github.event.pull_request.labels.*.name, 'skip changelog')
+    if: ${{ github.event_name == 'pull_request' }} && !contains(github.event.pull_request.labels.*.name, 'skip changelog')
     name: Check CHANGELOG.md for updates
     steps:
         - uses: actions/checkout@v3
