### 2023-08-11 #5756 @editorialbot
[REVIEW]: ZOSPy: optical ray tracing in Python through OpticStudio
**Submitting author:** @LucVV (<a href="http://orcid.org/0000-0001-8290-9071">Luc van Vught</a>)
**Repository:** https://github.com/MREYE-LUMC/ZOSPy
**Branch with paper.md** (empty if default branch): JOSS_submission
**Version:** v1.2.1
