### 2023-08-11 #5756 @plaplant
@brandondube you should be able to run the `@editorialbot generate my checklist` command in this thread now. Let me know if you run into issues. Thanks!
