### 2023-08-17 MREYE-LUMC/ZOSPy#35 @brandondube
Expanded documentation
This issue is part of my JOSS review, https://github.com/openjournals/joss-reviews/issues/5756

Copying verbatim, the grading criteria are:

### Documentation

Most of these do exist.  And the docstrings are even quite good, with the few I went looking for in the source code replete with examples!  However, this is a large software package, and when a user visits its website (the Github repo is all that exists for this, as best I can tell), perhaps just 5% or something like that of it "is documented" in the README.

I think it would be a great enhancement if a documentation site was set up for your package, for example using sphinx and hosted on readthedocs.  Or really whatever you like, but sphinx and the `nbsphinx` extension allow you to make quite delightful documentation with jupyter notebooks.
