### 2023-08-17 MREYE-LUMC/ZOSPy#36 @brandondube
JOSS paper content
This issue is part of my JOSS review, https://github.com/openjournals/joss-reviews/issues/5756

When I read your paper, the main on-its-face takeaway is that using the native or as-provided interface to Zemax leaves room for improvement, and the principal contribution of ZOSPy is a better/easier to use one.

For example, quoting

> With ZOSPy, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, we hope that the easy integration of the ray tracing workflow from optical design to analyses into Python scripts or Jupyter Notebooks facilitates open science.

By placing about a page of Python code as the first "what is it and how to use it" for the paper, that particular example in its whole is given quite a bit of significance.  In that example, the lens data manager is filled in programmatically, with

```
input_beam = oss.LDE.InsertNewSurfaceAt(1)
input_beam.Thickness = 10
# Make a 10 mm thick lens with a radius of curvature of 30mm
# and material type BK10
front_surface = oss.LDE.GetSurfaceAt(2)
front_surface.Radius = 30
front_surface.Thickness = 10
front_surface.SemiDiameter = 15
front_surface.Material = "BK10"
back_surface = oss.LDE.InsertNewSurfaceAt(3)
back_surface.Radius = -30
```

This would be equivalent to, for example with Code V MacroPlus
```
INS S1
THI S2 10
INS S2
RDY S2 30
THI S2 10
GLA S2 BK10
INS S3
RDY S3 -30
```

So, a takeaway that one is inclined to make is that the prescription is stored in python form.  The paper also criticizes the .zmx file format as proprietary, which re-inforces this idea.  It is not clear that there is an alternative offered by ZOSPy to this; a python script creating a prescription in a non-uniform way is not really a "file format."  So while the ZOSPy non-alternative is open source and cannot be proprietary by its nature, I don't think a solution to this problem has been presented, don't think it should be in the paper if a remedy to the problem is not present.

The paper also talks about "small number of" optical systems for the world without ZOSPy, which I find to be a strange point in two ways:

1) non-survey papers would concern themselves only with a small number of systems anyway (the ones being studied)

2) an irregular form of prescription entry would seem in conflict with analyzing large numbers of systems in an automated way

---

Claims are also made to ZOSPy being pythonic, but it seems most or all of the API is in CamelCase, which python only uses for class names, other variables being in snake_case.  So, for example, `oss.LDE.insert_new_surface_at(3)`, instead of the above.

---

Figure 2 is also not taken from the cited papers (I looked, having a copyright related concern about that), and is probably not generated easily using only ZOSPy?  Neither the paper nor the docs (README, related #35 ) show how to extract ray path data with ZOSPy.  I think the documentation should either cover how to generate similar graphics, or the paper be made more explicit about what portion of Fig 2 comes from ZOSPy, and what portion is made with other software.

---

---

Please do not misunderstand, I think something better than what Ansys provides directly is worth creating and worth publishing.  But I think that the software should stand on its own merits, and the criticisms made of alternatives in the paper should be directly addressed (remedied) by ZOSPy
