### 2023-08-20 MREYE-LUMC/ZOSPy#35 @LucVV
Thank you for the feedback, we will look into this and get back to you!
