+++ docs/api.md
@@ -1,0 +1,14 @@
+---
+title: API
+---
+
+# API documentation
+
+```{toctree}
+:maxdepth: 1
+:caption: Contents
+
+api/core
+api/analyses
+```
+
