+++ docs/examples.md
@@ -1,0 +1,16 @@
+---
+title: Examples flurp
+---
+
+## Examples
+
+Below are examples of various functionalities and applications of ZOSPy.
+
+```{toctree}
+:maxdepth: 1
+
+examples/thick_lens
+examples/polarization_prism
+examples/escudero_sanz
+```
+   
