+++ docs/api/analyses.md
@@ -1,0 +1,62 @@
+# Analyses
+
+```{eval-rst}
+.. automodule:: zospy.analyses
+   :members:
+```
+
+## MTF
+
+```{eval-rst}
+.. automodule:: zospy.analyses.mtf
+   :members:
+```
+
+## Polarization
+
+```{eval-rst}
+.. automodule:: zospy.analyses.polarization
+   :members:
+```
+
+## PSF
+
+```{eval-rst}
+.. automodule:: zospy.analyses.psf
+   :members:
+```
+
+## Rays and spots
+
+```{eval-rst}
+.. automodule:: zospy.analyses.raysandspots
+   :members:
+```
+
+## Reports
+
+```{eval-rst}
+.. automodule:: zospy.analyses.reports
+   :members:
+```
+
+## Surface
+
+```{eval-rst}
+.. automodule:: zospy.analyses.surface
+   :members:
+```
+
+## System viewers
+
+```{eval-rst}
+.. automodule:: zospy.analyses.systemviewers
+   :members:
+```
+
+## Wavefront
+
+```{eval-rst}
+.. automodule:: zospy.analyses.wavefront
+   :members:
+```
