+++ docs/compatibility.md
@@ -1,0 +1,25 @@
+# Compatibility
+
+```{include} compatibility_table.md
+```
+
+(compatibility/23-0-1)=
+## OpticStudio 23.0.1
+
+No known problems.
+
+(compatibility/20-3-2)=
+## OpticStudio 20.3.2
+
+:::{list-table}
+* - Type
+  - Python
+  - Description
+* - ℹ
+  - All
+  - The output of `zospy.analyses.polarization.TestPolarizationPupilMap` differs from the reference OpticStudio version in the returned Orientation. 
+    This is a difference between the two OpticStudio versions, not an issue with ZOSPy. 
+    As a result, the following unit tests fail:
+     - `test_polarization_pupil_map_matches_reference_data[1-0-0-0-Image-11x11]`
+     - `test_polarization_pupil_map_matches_reference_data[1-1-45-90-Image-17x17]`
+:::
