+++ docs/compatibility_table.md
@@ -1,0 +1,15 @@
+| OpticStudio |   20.3.2    |   23.0.1    |
+|-------------|:-----------:|:-----------:|
+| Python 3.9  | [ℹ][20.3.2] | [✔][23.0.1] |
+| Python 3.10 | [ℹ][20.3.2] | [✔][23.0.1] |
+| Python 3.11 | [ℹ][20.3.2] | [✔][23.0.1] |
+
+:::{dropdown} Table legend
+✔: This version works without problems.  
+ℹ: The output of all ZOSPy analyses matches the direct OpticStudio output, but the OpticStudio output has shown differences with the output of the reference version (currently **OpticStudio 23 R1.01**).  
+⚠: ZOSPy does not fully work.  
+❌: ZOSPy does not work at all.
+:::
+
+[23.0.1]: compatibility/23-0-1
+[20.3.2]: compatibility/20-3-2
