+++ docs/api/core.md
@@ -1,0 +1,6 @@
+# Core
+
+```{eval-rst}
+.. automodule:: zospy.zpcore
+   :members:
+```
