+++ docs/requirements.txt
@@ -1,1 +1,1 @@
-..[docs]
+.[docs]
