+++ pyproject.toml
@@ -41,6 +41,7 @@ docs = [
     "ipython",
     "myst-parser",
     "nbsphinx",
+    "nbsphinx_link",
     "numpydoc",
     "sphinx",
     "sphinx-book-theme",
