### 2023-08-31 3e66be3
Skip config when run by readthedocs
