+++ .readthedocs.yaml
@@ -11,6 +11,7 @@ build:
     python: "3.11"
   apt_packages:
     - pandoc
+    - dotnet-runtime-7.0
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
