### 2023-08-31 01496b3
Undo changes in config.py in favour of configuring a locale in readthedocs
