### 2023-08-31 789365c
Set locale and environment variables prior to build
