+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ build:
     - locales
   jobs:
     post_system_dependencies:
-      - locale-gen en_US.UTF-8
+      - su -c "locale-gen en_US.UTF-8"
     pre_build:
       - export PYTHONNET_RUNTIME="coreclr"
 
