### 2023-08-31 9537e32
Try to set the locale by installing a language pack
