### 2023-08-31 76e6093
Attempt to fake the existence of the winreg library
