+++ .readthedocs.yaml
@@ -14,14 +14,9 @@ build:
     - dotnet-runtime-7.0
     - language-pack-en-base
   jobs:
-#    post_system_dependencies:
-#      - su -c "locale-gen en_US.UTF-8"
     pre_build:
-      # Configure the .NET runtime
-      - export PYTHONNET_RUNTIME="coreclr"
-      # Set the locale to en_US, otherwise zospy.api.config will crash
-#      - export LC_ALL="en_US.UTF-8"
-#      - export LC_CTYPE="en_US.UTF-8"
+      # Fake the existence of winreg
+      - touch winreg.py
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
