### 2023-08-31 2a35455
A more sophisticated attempt to fake the existence of the winreg library
