+++ .readthedocs.yaml
@@ -16,7 +16,20 @@ build:
   jobs:
     pre_build:
       # Fake the existence of winreg
-      - touch winreg.py
+      - mkdir winreg
+      - mkdir winreg/winreg
+      - touch winreg/winreg/__init__.py
+      - |
+          printf \
+          '[build-system]
+          requires = ["flit_core >=3.2,<4"]
+          build-backend = "flit_core.buildapi"
+          
+          [project]
+          name = "winreg"
+          description = "a fake version of winreg"
+          version = "1.0.0"\n' > winreg/pyproject.toml
+      - pip install ./winreg
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
