### 2023-08-31 af210bd
Debug building winreg
