+++ .readthedocs.yaml
@@ -16,9 +16,9 @@ build:
   jobs:
     pre_build:
       # Fake the existence of winreg
-      - mkdir winreg
-      - mkdir winreg/winreg
-      - touch winreg/winreg/__init__.py
+      - mkdir ./winreg
+      - mkdir ./winreg/winreg
+      - touch ./winreg/winreg/__init__.py
       - |
           printf \
           '[build-system]
@@ -28,7 +28,8 @@ build:
           [project]
           name = "winreg"
           description = "a fake version of winreg"
-          version = "1.0.0"\n' > winreg/pyproject.toml
+          version = "1.0.0"\n' > ./winreg/pyproject.toml
+      - cat ./winreg/pyproject.toml
       - pip install ./winreg
 
 # Build documentation in the "docs/" directory with Sphinx
