### 2023-08-31 50aae43
Ensure full command is executed
