### 2023-08-31 77d39af
Add documentation for solvers
