### 2023-08-31 7b361bc
Another try to fake winreg
