+++ .readthedocs.yaml
@@ -19,14 +19,14 @@ build:
       - mkdir ./winreg
       - mkdir ./winreg/winreg
       - touch ./winreg/winreg/__init__.py
-      - |
-          printf \
-          '[build-system]\n \
-          requires = ["flit_core >=3.2,<4"]\n \
-          build-backend = "flit_core.buildapi"\n \
-          [project]\n \
-          name = "winreg"\n \
-          description = "a fake version of winreg"\n \
+#      - printf '[build-system]\nrequires = ["flit_core >=3.2,<4"]\nbuild-backend = "flit_core.buildapi"\n[project]\nname = "winreg"\ndescription = "a fake version of winreg"\nversion = "1.0.0"\n' > ./winreg/pyproject.toml
+      - >-
+          printf '[build-system]\n
+          requires = ["flit_core >=3.2,<4"]\n
+          build-backend = "flit_core.buildapi"\n\n
+          [project]\n
+          name = "winreg"\n
+          description = "a fake version of winreg"\n
           version = "1.0.0"\n' > ./winreg/pyproject.toml
       - cat ./winreg/pyproject.toml
       - pip install ./winreg
