### 2023-08-31 20cb26e
Dynamically create an empty winreg module in docs/conf.py
