+++ docs/conf.py
@@ -3,6 +3,11 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+import os
+import sys
+
+from types import ModuleType
+
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
@@ -36,3 +41,8 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 html_theme = "sphinx_book_theme"
 html_static_path = ["_static"]
 html_theme_options = {"home_page_in_toc": True}
+
+# -- Enviroment configuration ------------------------------------------------
+if os.environ.get("READTHEDOCS"):
+    winreg_module = ModuleType("winreg")
+    sys.modules["winreg"] = winreg_module
