### 2023-09-01 eae0142
Fix most build problems by using autodoc options
