+++ docs/api/core.md
@@ -1,6 +1,0 @@
-# Core
-
-```{eval-rst}
-.. automodule:: zospy.zpcore
-   :members:
-```
