+++ docs/_templates/custom-class.rst
@@ -1,0 +1,9 @@
+{{ fullname | escape | underline}}
+
+.. currentmodule:: {{ module }}
+
+.. autoclass:: {{ objname }}
+   :members:
+   :show-inheritance:
+   :no-inherited-members:
+

[…]
