+++ zospy/solvers/__init__.py
@@ -1,6 +1,4 @@
-"""
-OpticStudio Solvers
-===================
+"""OpticStudio Solvers.
 
 `zospy.solvers` provides easy access to various solvers offered by OpticStudio.
 Solvers have the general structure `solver_function(cell, *args)`, with `cell` the EditorCell of the solver,
@@ -8,7 +6,6 @@ followed by the parameters for the solver.
 
 Examples
 --------
-
 Make the radius of a surface variable:
 
 >>> surface = oss.InsertNewSurfaceAt(1)
