### 2023-09-04 320302c
Add a documentation website on ReadTheDocs (#37)
* Add documentation

* Add requirements.txt for readthedocs

* Add .readthedocs.yaml

* Add nbsphinx_link to documentation dependencies

* Remove GitHub pages configuration

* Add documentation badge to README

* Attempt to fake the existence of the winreg library

* Configure autodoc to mock some libraries

* Improve automatic generation of API documentation
