+++ zospy/zpcore.py
@@ -16,7 +16,7 @@ class OpticStudioSystem:
     """Wrapper for OpticStudio System instances."""
 
     def __init__(self, zos_instance, system_instance):
-        """Initiates the OpticStudioSystem.
+        """Initiate the OpticStudioSystem.
 
         Parameters
         ----------
@@ -32,6 +32,7 @@ class OpticStudioSystem:
 
     @property
     def SystemName(self) -> str:
+        """Name of the current optical system."""
         return self._System.SystemName
 
     @SystemName.setter
@@ -40,105 +41,156 @@ class OpticStudioSystem:
 
     @property
     def SystemID(self) -> int:
+        """Unique identifier for the optical system.
+
+        This identifier can be used to differentiate between multiple `OpticStudioSystem` instances.
+
+        Examples
+        --------
+        Create two different optical systems in a single instance of OpticStudio:
+
+        >>> import zospy as zp
+        >>> zos = zp.ZOS()
+        >>> oss1 = zos.connect_as_standalone(return_primary_system=True)
+        >>> oss2 = zos.create_new_system()
+        >>> oss1.SystemID != oss2.SystemID
+        True
+        """
         return self._System.SystemID
 
     @property
     def Mode(self) -> str:
+        """Mode of the optical system. Either "Sequential" or "NonSequential"."""
         return str(self._System.Mode)
 
     @property
     def SystemFile(self) -> str:
+        """File path to the current optical system."""
         return self._System.SystemFile
 
     @property
     def IsNonAxial(self) -> bool:
+        """Indicates whether the optical system is axial and sequential.
+
+        `True` if the system is non-axial, `False` otherwise.
+        """
         return self._System.IsNonAxial
 
     @property
     def NeedsSave(self) -> bool:
+        """Indicates if the optical system contains unsaved changes."""
         return self._System.NeedsSave
 
     @property
     def SystemData(self) -> _ZOSAPI.SystemData.ISystemData:
+        """Data for configuring everything in the system explorer.
+
+        Examples
+        --------
+        Change the aperture type to "FloatByStopSize":
+
+        >>> oss.SystemData.Aperture.ApertureType = zp.constants.SystemData.ZemaxApertureType.FloatByStopSize
+
+        Add a wavelength of 543 nm with weight 1:
+
+        >>> oss.SystemData.Wavelengths.AddWavelength(0.543, 1)
+        """
         return self._System.SystemData
 
     @property
     def LDE(self) -> _ZOSAPI.Editors.LDE.ILensDataEditor:
+        """Lens Data Editor."""
         return self._System.LDE
 
     @property
     def NCE(self) -> _ZOSAPI.Editors.NCE.INonSeqEditor:
+        """Non-Sequential Component Editor."""
         return self._System.NCE
 
     @property
     def MFE(self) -> _ZOSAPI.Editors.MFE.IMeritFunctionEditor:
+        """Merit Function Editor."""
         return self._System.MFE
 
     @property
     def TDE(self) -> _ZOSAPI.Editors.TDE.IToleranceDataEditor:
+        """Tolerance Data Editor."""
         return self._System.TDE
 
     @property
     def MCE(self) -> _ZOSAPI.Editors.MCE.IMultiConfigEditor:
+        """Multi-Configuration Editor."""
         return self._System.MCE
 
     @property
     def Analyses(self) -> _ZOSAPI.Analysis.I_Analyses:
+        """Analyses for the current system."""
         return self._System.Analyses
 
     @property
     def Tools(self) -> _ZOSAPI.Tools.IOpticalSystemTools:
+        """Interface to run various tools on the optical system."""
         return self._System.Tools
 
     @property
[…]
