+++ zospy/functions/__init__.py
@@ -1,3 +1,12 @@
+"""
+Utility functions for ZOSPy.
+
+`zospy.functions` contains utility functions for zospy. These functions are available through its submodules:
+
+- **`zospy.functions.lde`** provides helper functions for the Lens Data Editor (LDE);
+- **`zospy.functions.nce`** provides helper functions for the Non-sequential Component Editor (NCE).
+"""
+
 from . import lde, nce
 
 __all__ = ("lde", "nce")
