### 2023-09-09 MREYE-LUMC/ZOSPy#35 @LucVV
Hi @brandondube, we have addressed this feedback by hosting the documentation on [readthedocs](https://zospy.readthedocs.io/en/latest/ ) as suggested.
