+++ paper/paper.md
@@ -37,7 +37,7 @@ Ray tracing simulations are widely used to design, optimize and analyze optical 
 
 Although OpticStudio offers an API, the `ZOS-API`, using this API in Python is complex and time-consuming. It involves, for example, establishing a connection with the API through the .NET framework, casting between .NET and Python datatypes, identifying which constants need to be set in specific cases, and working around  non-uniform methods of parsing the output [@GettingStartedWithZOSAPI]. This leads to studies which, in practice, largely rely on user interaction and therefore generally only assess a small number of optical systems. It furthermore hinders open science initiatives as sharing of methods is limited to screenshots [@Hong2011], simple tables [@Naeem2022] or optical systems saved in the proprietary file format of OpticStudio [@Polans2018]. 
 
-With `ZOSPy`, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, we hope that the easy integration of the ray tracing workflow from optical design to analyses into Python scripts or Jupyter Notebooks facilitates open science. 
+With `ZOSPy`, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, as users can directly share their optical system via Python scripts or Jupyter Notebooks rather than screenshots or optical systems saved in proprietary file formats, we strive to facilitate open science. 
 
 # Functionality
 `ZOSPy` is, in its most basic form, a Python wrapper around the OpticStudio API. It facilitates the .NET connection required to connect to OpticStudio through its API, as well as all subsequent casting of variables between .NET and Python. Additionally, it provides object-oriented methods to define surfaces and their optical properties. Furthermore, it offers single-line, easy to understand, methods to perform analyses that return the analysis results in a uniform way. As a result, `ZOSPy` enables a straight-forward interaction with OpticStudio and improves code readability, which facilitates method sharing between scientists.
