+++ paper/paper.bib
@@ -23,6 +23,30 @@
   doi = {10.1364/BOE.2.001649}
 }
 
+@article{Ellis2001,
+  title = {Sharp-edged intraocular lens design as a cause of permanent glare},
+  author = {Ellis, Mark F},
+  journal = {Journal of Cataract \& Refractive Surgery},
+  volume = {27},
+  number = {7},
+  pages = {1061--1064},
+  year = {2001},
+  publisher = {Elsevier},
+  doi = {10.1016/S0886-3350(00)00856-7}
+}
+
+@article{Holladay1999,
+  title = {Analysis of edge glare phenomena in intraocular lens edge designs},
+  author = {Holladay, Jack T and Lang, Alan and Portney, Val},
+  journal = {Journal of Cataract \& Refractive Surgery},
+  volume = {25},
+  number = {6},
+  pages = {748--752},
+  year = {1999},
+  publisher = {Elsevier},
+  doi = {10.1016/S0886-3350(99)00038-3}
+}
+
 @article{Hong2011,
   title = {Ray-tracing optical modeling of negative dysphotopsia},
   author = {Hong, Xin and Liu, Yueai and Karakelle, Mutlu and Masket, Samuel and Fram, Nicole R},
@@ -58,6 +82,18 @@
   url = {https://www.mdpi.com/2673-3269/3/3/28},
   issn = {2673-3269},
   doi = {10.3390/opt3030028}
+}
+
+@article{Osher2008,
+  title = {Negative dysphotopsia: long-term study and possible explanation for transient symptoms},
+  author = {Osher, Robert H},
+  journal = {Journal of Cataract \& Refractive Surgery},
+  volume = {34},
+  number = {10},
+  pages = {1699--1707},
+  year = {2008},
+  publisher = {Elsevier},
+  doi = {10.1016/j.jcrs.2008.06.026}
 }
 
 @article{Polans2018,
