### 2023-09-13 6bce3db
Updated use cases section and figure 2
