### 2023-09-13 cfa459b
Minor change to figure caption for figure 2
