### 2023-09-14 MREYE-LUMC/ZOSPy#36 @brandondube
>The aim of ZOSPy is not provide a new file format, but rather to provide a way of documenting the entire process from creating an optical system, defining the settings of analyses and performing these analyses. This way, the entire process can be shared in an insightful manner. We have altered the last paragraph of the statement of need to clarify this:

>>With ZOSPy, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, as users can directly share their Python scripts or Jupyter Notebooks rather than screenshots or optical systems saved in proprietary file formats, we strive to facilitate open science.

I guess it is not clear to me that this is something "new" with this package.  If there were no python API at all, one could share a .ZMX file, and .ZPL macros for analysis.  One may feel (including myself) that Python > ZPL, and so there is a benefit.  And especially that ZOSPy > ZOS-API.  But I don't think the "sharing" part is new, moreso that scripts, etc, shared are less noisy / easier to read and comprehend.  The closing paragraph, which criticizes sharing alternatives in _proprietary_ formats I still feel is not a real difference; so long as you are wed to OpticsStudio (proprietary), then the script cannot be used without proprietary software anyway, and whether the optical design was in a proprietary .ZMX file or otherwise is rather moot.

Perhaps, as an example, this could be edited to emphasize legibility

>>With ZOSPy, we provide an easy to use and accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding.  Colleagues who are not familiar with the intricacies of the `ZOS-API` interface are able to read and comprehend scripts that use ZOSPy, providing greater accessibility to analysis conducting with ZOSPy, compared to directly using `ZOS-API`.

Just a thought

--

> ...Where the clinical studies in this area often evaluate over 1000 different eyes, their ray-tracing counter parts are limited to only a small number of patients. We have clarified this in the text:

I guess I do not really see, or agree with the alternatives being contrasted here.  Zemax can perform Monte Carlo analysis, in which many (100s) of perturbations of an optical system can be generated and analyzed in an automated way.  When the systems can be described in the same set of surfaces, but different parameters, the use case you describe seems to map quite neatly to the MC functionality in Zemax, modulo that you do not desire randomness.  To me, it seems like a strange angle to contrast your software by implying that Zemax can't do analysis of large ranges of optical systems.  This is a fairly core part of designing something through all the way to fabrication, where MC or something like it is done to derive and verify the tolerances.

I am well satisfied with the other changes

Thanks!
