+++ paper/paper.bib
@@ -1,4 +1,16 @@
 
+@article{Alfonso2007,
+  title={Prospective visual evaluation of apodized diffractive intraocular lenses},
+  author={Alfonso, Jos{\'e} F and Fern{\'a}ndez-Vega, Luis and Baamonde, M Bego{\~n}a and Mont{\'e}s-Mic{\'o}, Robert},
+  journal={Journal of Cataract \& Refractive Surgery},
+  volume={33},
+  number={7},
+  pages={1235--1243},
+  year={2007},
+  publisher={Elsevier},
+  doi = {10.1016/j.jcrs.2007.03.034}
+}
+
 @article{Artal2023,
   title = {Inverted meniscus intraocular lens as a better optical surrogate of the crystalline lens},
   author = {Artal, Pablo and Ginis, Harilaos and Christaras, Dimitrios and Villegas, Eloy A and Tabernero, Juan and Prieto, Pedro M},
@@ -84,18 +96,6 @@
   doi = {10.3390/opt3030028}
 }
 
-@article{Osher2008,
-  title = {Negative dysphotopsia: long-term study and possible explanation for transient symptoms},
-  author = {Osher, Robert H},
-  journal = {Journal of Cataract \& Refractive Surgery},
-  volume = {34},
-  number = {10},
-  pages = {1699--1707},
-  year = {2008},
-  publisher = {Elsevier},
-  doi = {10.1016/j.jcrs.2008.06.026}
-}
-
 @article{Polans2018,
   author = {James Polans and Bart Jaeken and Ryan P. McNabb and Pablo Artal and Joseph A. Izatt},
   journal = {Optica},
@@ -121,6 +121,18 @@
   year = {2020},
   publisher = {LWW},
   doi = {10.1097/j.jcrs.0000000000000103}
+}
+
+@article{vanVught2020,
+  title={Distinct differences in anterior chamber configuration and peripheral aberrations in negative dysphotopsia},
+  author={van Vught, Luc and Luyten, Gregorius PM and Beenakker, Jan-Willem M},
+  journal={Journal of Cataract and Refractive Surgery},
+  volume={46},
+  number={7},
+  pages={1007},
+  year={2020},
+  publisher={Wolters Kluwer Health},
+  doi = {10.1097/j.jcrs.0000000000000206}
 }
 
 @article{vanVught2022,
