### 2023-09-15 9bc4d18
Updated paper following latest feedback
