+++ paper/paper.md
@@ -35,9 +35,9 @@ Zemax OpticStudio (Ansys, Inc) is a commonly used software package for designing
 # Statement of need
 Ray tracing simulations are widely used to design, optimize and analyze optical systems. Its applications are diverse, ranging from designing spectrometers [@Naeem2022] or telescopes [@Zhang2023], to understanding the optics of the human eye [@Simpson2020; @vanVught2022]. Moreover, in ophthalmology, ray tracing is used to optimize the outcomes of cataract surgery [@Canovas2011; @Artal2023] and evaluate the accuracy of ocular radiotherapy [@Jaarsma2023]. These optical simulations are often performed in OpticStudio, which offers a powerful set of tools to design, optimize and evaluate optical systems.
 
-Although OpticStudio offers an API, the `ZOS-API`, using this API in Python is complex and time-consuming. It involves, for example, establishing a connection with the API through the .NET framework, casting between .NET and Python datatypes, identifying which constants need to be set in specific cases, and working around  non-uniform methods of parsing the output [@GettingStartedWithZOSAPI]. This leads to studies which, in practice, largely rely on user interaction and therefore mostly assess a small number of optical systems. Although this is generally sufficient for the design of an optical system, other applications would benefit from the evaluation a large set of systems. In vision science, for example, clinical studies typically analyze vision-related complaints in cohorts of hundreds of eyes [@Ellis2001; @Osher2008], but the ray-tracing studies aiming to link these complaints to the subject’s ocular optics are limited to a small number of eyes [@Holladay1999; @Simpson2020]. Furthermore, the sharing of methodology is often limited to screenshots [@Hong2011], simple tables [@Naeem2022] or optical systems saved in the proprietary file format of OpticStudio [@Polans2018], which hinders open science initiatives. 
+Although OpticStudio offers an API, the `ZOS-API`, using this API in Python is complex and time-consuming. It involves, for example, establishing a connection with the API through the .NET framework, casting between .NET and Python datatypes, identifying which constants need to be set in specific cases, and working around  non-uniform methods of parsing the output [@GettingStartedWithZOSAPI]. This leads to studies which, in practice, largely rely on user interaction. Although OpticStudio can perform Monte Carlo analyses, where a large number of random perturbations of the system are generated and analysed in an automated way, this type of automation is not suitable when large sets of specific, non-random, combinations of parameters need to be analysed. In vision science, for example, ray tracing is used to design artificial lenses for the eye [@Ellis2001; @Holladay1999], but their evaluation in a large set of patients is hindered as the anatomical parameters of each subject’s eye need to be entered manually. As a result, clinical studies typically describe vision-related complaints in cohorts of hundreds of eyes [@Alfonso2007], but the ray tracing studies aiming to link these outcomes to the subject’s ocular optics are limited to a small number of eyes [@Simpson2020; @vanVught2020].
 
-With `ZOSPy`, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, as users can directly share their optical system via Python scripts or Jupyter Notebooks rather than screenshots or optical systems saved in proprietary file formats, we strive to facilitate open science. 
[…]
