### 2023-09-15 9f24e2d
Small fix for reference formatting
