+++ paper/paper.bib
@@ -125,7 +125,7 @@
 
 @article{vanVught2020,
   title={Distinct differences in anterior chamber configuration and peripheral aberrations in negative dysphotopsia},
-  author={van Vught, Luc and Luyten, Gregorius PM and Beenakker, Jan-Willem M},
+  author={{van Vught}, Luc and Luyten, Gregorius PM and Beenakker, Jan-Willem M},
   journal={Journal of Cataract and Refractive Surgery},
   volume={46},
   number={7},
@@ -137,7 +137,7 @@
 
 @article{vanVught2022,
   title = {Effect of anatomical differences and intraocular lens design on Negative Dysphotopsia.},
-  author = {van Vught, Luc and Que, Ivo and Luyten, Gregorius PM and Beenakker, Jan-Willem M},
+  author = {{van Vught}, Luc and Que, Ivo and Luyten, Gregorius PM and Beenakker, Jan-Willem M},
   journal = {Journal of Cataract \& Refractive Surgery},
   pages = {10--1097},
   year = {2022},
