### 2023-09-15 MREYE-LUMC/ZOSPy#36 @LucVV
Hi @brandondube, thank you for the swift reply. We are happy to hear that you are satisfied with most of our changes! We have made some additional changes following your latest feedback:

---
> I guess it is not clear to me that this is something "new" with this package. If there were no python API at all, one could share a .ZMX file, and .ZPL macros for analysis. One may feel (including myself) that Python > ZPL, and so there is a benefit. And especially that ZOSPy > ZOS-API. But I don't think the "sharing" part is new, moreso that scripts, etc, shared are less noisy / easier to read and comprehend. The closing paragraph, which criticizes sharing alternatives in proprietary formats I still feel is not a real difference; so long as you are wed to OpticsStudio (proprietary), then the script cannot be used without proprietary software anyway, and whether the optical design was in a proprietary .ZMX file or otherwise is rather moot.
>
> Perhaps, as an example, this could be edited to emphasize legibility
>
>> With ZOSPy, we provide an easy to use and accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Colleagues who are not familiar with the intricacies of the ZOS-API interface are able to read and comprehend scripts that use ZOSPy, providing greater accessibility to analysis conducting with ZOSPy, compared to directly using ZOS-API.
>
> Just a thought

We have changed this section as proposed:

_With ZOSPy, we provide an easy-to-use and accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. As a result, those who are not familiar with the intricacies of the `ZOS-API` interface will be able to read and comprehend scripts that use `ZOSPy`. Thereby, `ZOSPy`  provides greater accessibility to conducting analyses in OpticStudio through Python than directly using the `ZOS-API`._

---

> It may help de-emphasize this (which i think is not your central point) to just show fig 1, and leave the minimum working example code to an example file, which can be referenced in the paper's text

We agree that the entire example code to much for this paper. However, we feel that showing some of the code shows the ease-of-use and comprehensibility of the resulting code. We have therefore shorted the example code to only show some essential lines of code and have referred to the example script for the remainder of the code:

> These examples provide new users with an easy start with `ZOSPy`. Part of a [simple example](https://zospy.readthedocs.io/en/latest/examples/thick_lens.html) of using `ZOSPy` to create and evaluate a thick lens is shown below, and the corresponding results are shown in \autoref{fig:1}.
>
> ```python
> #...
>
> # Make a 10 mm thick lens with a radius of curvature of 30 mm
> # and material type BK10
> front_surface = oss.LDE.GetSurfaceAt(2)
> front_surface.Radius = 30
> front_surface.Thickness = 10
> front_surface.SemiDiameter = 15
> front_surface.Material = "BK10"
>
> back_surface = oss.LDE.InsertNewSurfaceAt(3)
> back_surface.Radius = -30
> back_surface.Thickness = 29
> back_surface.SemiDiameter = 15
>
> #...
>
> # Render the model
> draw3d = zp.analyses.systemviewers.viewer_3d(oss)
>
> # Analyze the system by calculating the Modulation Transfer Function (MTF)
> # and the Point Spread Function (PSF)
> mtf = zp.analyses.mtf.fft_through_focus_mtf(oss, sampling="512x512", deltafocus=2.5, numberofsteps=51)
>
> huygens_psf = zp.analyses.psf.huygens_psf(oss, pupil_sampling="512x512", image_sampling="512x512", normalize=True)
> ```

---
> I guess I do not really see, or agree with the alternatives being contrasted here. Zemax can perform Monte Carlo analysis, in which many (100s) of perturbations of an optical system can be generated and analyzed in an automated way. When the systems can be described in the same set of surfaces, but different parameters, the use case you describe seems to map quite neatly to the MC functionality in Zemax, modulo that you do not desire randomness. To me, it seems like a strange angle to contrast your software by implying that Zemax can't do analysis of large ranges of optical systems. This is a fairly core part of designing something through all the way to fabrication, where MC or something like it is done to derive and verify the tolerances.

we understand and agree with you point. We have therefore rephrased these sentences to better explain the difference between Monte Carlo type of analyses which are possible, and the non-random permutations required for some medically oriented studies:

_Although OpticStudio offers an API, the `ZOS-API`, using this API in Python is complex and time-consuming. It involves, for example, establishing a connection with the API through the .NET framework, casting between .NET and Python datatypes, identifying which constants need to be set in specific cases, and working around  non-uniform methods of parsing the output [@GettingStartedWithZOSAPI]. This leads to studies which, in practice, largely rely on user interaction. Although OpticStudio can perform Monte Carlo analyses, where a large number of random perturbations of the system are generated and analysed in an automated way, this type of automation is not suitable when large sets of specific, non-random, combinations of parameters need to be analysed. In vision science, for example, ray tracing is used to design artificial lenses for the eye [@Ellis2001; @Holladay1999], but their evaluation in a large set of patients is hindered as the anatomical parameters of each subject’s eye need to be entered manually. As a result, clinical studies typically describe vision-related complaints in cohorts of hundreds of eyes [@Alfonso2007], but the ray tracing studies aiming to link these outcomes to the subject’s ocular optics are limited to a small number of eyes [@Simpson2020; @vanVught2020]._
