### 2023-09-18 56abc2b
Automatically copy examples and insert banner above each example
