### 2023-09-18 00f637b
Remove nbsphinx_link extension and commented out extensions
nbsphinx_link is not necessary anymore, because the examples are  copied from the project root.
