### 2023-09-21 f3cd3aa
Update compatibility information with version 23 R2.1
