+++ tests/test_zpcore.py
@@ -1,8 +1,10 @@
+import locale
 from types import SimpleNamespace
 
 import pytest
 
 import zospy as zp
+from zospy import constants
 
 
 def patch_zos(zos: zp.ZOS, monkeypatch: pytest.MonkeyPatch):
@@ -77,6 +79,22 @@ def test_create_simple_system(simple_system):
     assert simple_system.LDE.NumberOfSurfaces == 5
 
 
+@pytest.fixture()
+@pytest.mark.require_mode("standalone")
+def oss_with_modifiable_config(zos: zp.ZOS, connection_mode, tmp_path) -> zp.zpcore.OpticStudioSystem:
+    config_file = tmp_path / "opticstudio_config_file.CFG"
+    zos.Connection.PreferencesFile = str(config_file)
+    connected = zos.create_new_application()
+
+    oss = zos.get_primary_system()
+
+    yield oss
+
+    # Close the system
+    if connected:
+        zos.Application.CloseApplication()
+
+
 def test_version(optic_studio_version):
     assert optic_studio_version
 
@@ -115,3 +133,40 @@ def test_create_new_system(zos, oss, connection_mode):
 def test_create_new_system_raises_valueerror(zos, simple_system, connection_mode):
     with pytest.raises(ValueError):
         zos.create_new_system()
+
+
+class TestTxtFileEncoding:
+    @pytest.mark.parametrize(
+        "txtfile_encoding,expected_encoding", [("Unicode", "UTF-16-le"), ("ANSI", "LocalePreferredEncoding")]
+    )
+    def test_get_txtfile_encoding_returns_correct_result(
+        self, oss_with_modifiable_config, txtfile_encoding, expected_encoding, monkeypatch: pytest.MonkeyPatch
+    ):
+        def getpreferredencoding(**kwargs):
+            return "LocalePreferredEncoding"
+
+        monkeypatch.setattr(locale, "getpreferredencoding", getpreferredencoding)
+
+        oss_with_modifiable_config._ZOS.Application.Preferences.General.TXTFileEncoding = getattr(
+            constants.Preferences.EncodingType, txtfile_encoding
+        )
+
+        returned_encoding = oss_with_modifiable_config._ZOS.get_txtfile_encoding()
+
+        assert returned_encoding == expected_encoding
+
+    @pytest.mark.parametrize("txtfile_encoding", ["Unicode", "ANSI"])
+    def test_analysis_result_parsed_with_correct_encoding(self, oss_with_modifiable_config, txtfile_encoding, tmp_path):
+        oss_with_modifiable_config._ZOS.Application.Preferences.General.TXTFileEncoding = getattr(
+            constants.Preferences.EncodingType, txtfile_encoding
+        )
+
+        analysis = zp.analyses.base.new_analysis(
+            oss_with_modifiable_config, zp.constants.Analysis.AnalysisIDM.SystemData
+        )
+        analysis.ApplyAndWaitForCompletion()
+        txtoutfile = str(tmp_path / "test_analysis_result_parsed_with_correct_encoding.txt")
+        analysis.Results.GetTextFile(txtoutfile)
+
+        with open(txtoutfile, "r", encoding=oss_with_modifiable_config._ZOS.get_txtfile_encoding()) as txtfile:
+            assert "System/Prescription Data" in txtfile.readline()
