### 2023-09-25 fa4ca61
Fix LensUpdateMode setter (#40)
* Fixed bug in LensUpdateMode setter: use the correct ZOS-API SystemData property

* Updated CHANGELOG.md

---------

Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>
