### 2023-09-25 867049c
Consistent style throughout CHANGELOD.md
