### 2023-09-25 72badb0
Remove Python 3.8 from test labels
