+++ tox.ini
@@ -2,8 +2,8 @@
 min_version = 4.0
 
 labels =
-    test-extension = py3{8,9,10,11}-extension
-    test-standalone = py3{8,9,10,11}-standalone
+    test-extension = py3{9,10,11}-extension
+    test-standalone = py3{9,10,11}-standalone
     extension = format,lint,py3{9,10,11}-extension
 
 # Default environments that are run without further specification
