+++ docs/compatibility_table.md
@@ -1,8 +1,8 @@
-| OpticStudio |   20.3.2    |   23.0.1    |
-|-------------|:-----------:|:-----------:|
-| Python 3.9  | [ℹ][20.3.2] | [✔][23.0.1] |
-| Python 3.10 | [ℹ][20.3.2] | [✔][23.0.1] |
-| Python 3.11 | [ℹ][20.3.2] | [✔][23.0.1] |
+| OpticStudio |   20.3.2    |   23.0.1    | 23.2.1       |
+|-------------|:-----------:|:-----------:|:------------:|
+| Python 3.9  | [ℹ][20.3.2] | [✔][23.0.1] | [✔][23.2.1] |
+| Python 3.10 | [ℹ][20.3.2] | [✔][23.0.1] | [✔][23.2.1] |
+| Python 3.11 | [ℹ][20.3.2] | [✔][23.0.1] | [✔][23.2.1] |
 
 :::{dropdown} Table legend
 ✔: This version works without problems.  
@@ -11,5 +11,6 @@
 ❌: ZOSPy does not work at all.
 :::
 
+[23.2.1]: compatibility/23-2-1
 [23.0.1]: compatibility/23-0-1
 [20.3.2]: compatibility/20-3-2
