### 2023-09-25 9df7164
Synchronize with LUMC GitLab (#42)
Fixed:

* Erroneous parsing of analyses results when textfile encoding was not set to Unicode by implementing zospy.zpcore.ZOS.get_txt_file_encoding (!36)

Changed:

* Updated how and when constants in zospy.api.config are determined for more clarity (!39)
