+++ zospy/__init__.py
@@ -4,13 +4,14 @@ __version__ = "1.1.0"
 
 import logging
 
-import zospy.api.config as _config
+import zospy.api.config as config
 from zospy import analyses, functions, solvers
 from zospy.api import constants
 from zospy.zpcore import ZOS
 
+config.set_decimal_point_and_thousands_separator()
+
 __all__ = (
-    "_config",
     "analyses",
     "constants",
     "functions",
