### 2023-09-25 f69ea60
Delete .gitlab/merge_request_templates directory
