+++ .gitlab/merge_request_templates/default.md
@@ -1,13 +1,0 @@
-## Description
-
-Motivate your changes and describe what has been changed.
-
-## Related issues
-
-List any issues that are related to these changes.
-
-## Checklist
-
-- [ ] I added unit tests for new features or updated existing unit tests
-- [ ] All tests pass (if not, please explain which tests fail and why you didn't fix this before requesting a review)
-- [ ] I updated `CHANGELOG.md` with these changes
