### 2023-09-28 ea4379d
Add documentation option to PR template
