### 2023-09-28 5274864
Fix incorrect API link and extend explanation about OpticStudio boilerplate code
