### 2023-09-28 c1233c5
Get the version of ZOSPy without importing it
