### 2023-09-28 32edb67
Import importlib.metadata instead of only its version function
The 'version' variable also has a different meaning in conf.py, which causes the build to fail
