### 2023-09-28 75784f4
Add a banner with download links to examples (#46)
* Automatically copy examples and insert banner above each example

* Insert link to ZOSPy version for which the docs are build

* Remove nbsphinx_link extension and commented out extensions

nbsphinx_link is not necessary anymore, because the examples are  copied from the project root.

* Add repository link to header

* Remove empty cell from example notebook
