+++ docs/examples.md
@@ -6,11 +6,15 @@ title: Examples flurp
 
 Below are examples of various functionalities and applications of ZOSPy.
 
+<!-- 
+List the paths of the examples relative to the project root.
+The example folder is copied to the documentation folder before building the documentation.
+-->
 ```{toctree}
 :maxdepth: 1
 
-examples/thick_lens
-examples/polarization_prism
-examples/escudero_sanz
+examples/Simple thick lens/Simple thick lens
+examples/Escudero-Sanz eye model/Escudero-Sanz eye model
+examples/Polarization Prism/polarization_prism_example
 ```
    
