+++ docs/examples/README.md
@@ -1,12 +1,0 @@
-# Example: Simple thick lens
-
-This example shows how to create and analyze a simple thick lens as described in the paper [ZOSPy: optical ray tracing in Python through OpticStudio](../../Paper). The code resembles the example of the paper with the addition of plotting the analysis results.
-
-## Included functionalities
-* _Sequential mode_:
-  - Building a simple sequential optical system.
-  - Usage of `zospy.analyses.mtf.fft_through_focus_mtf` to calculate the FFT through focus MTF.
-  - Usage of `zospy.analyses.psf.huygens_psf()` to perform a Huygens PSF analysis.
-
-## Warranty and liability
-The examples are provided 'as is'. There is no warranty and rights cannot be derived from them, as is also stated in the general license of this repository.
