### 2023-11-15 7d3d080
Create bug_report.yml
