### 2023-11-16 #5756 @plaplant
@LucVV apolgies for the delays, and thanks for your patience! It looks like both of the reviewers have started working their ways through the reviews.

@brandondube thanks very much for your review! I see that you've started an issue with the content of the review (https://github.com/MREYE-LUMC/ZOSPy/issues/36). If you feel that the authors have addressed all the issues, we can close that issue and update the review checklist here to reflect that.

@alvesjnr thanks for filling out the checklist! It's generally helpful to open an issue (or series of issues) on the upstream repository outlining specific changes you would like to see. That way, the authors can address them and have a conversation about the suggested edits. Let me know if you have any questions about the process!
