### 2023-11-19 MREYE-LUMC/ZOSPy#36 @brandondube
Apologies, looks like I missed your reply.  I am satisfied with the changes, thank you for your work and willingness to modify the paper!
