### 2023-12-06 2d87ec4
Extend some descriptions in the issue form
