### 2023-12-11 d2b2597
Reformatting for ruff lint.
