### 2023-12-11 785015a
synchronize GitLab with GitHub (datagrid bugfix)
Co-authored-by: LucVV <l.van_vught@lumc.nl>
Co-authored-by: jwmbeenakker <j.w.m.beenakker@lumc.nl>
Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>
