+++ tests/analyses/test_surface.py
@@ -1,9 +1,8 @@
 import numpy as np
 import pytest
+from pandas.testing import assert_frame_equal
 
 from zospy.analyses.surface import curvature
-
-from pandas.testing import assert_frame_equal
 
 
 class TestCurvature:
