+++ tests/analyses/test_psf.py
@@ -1,6 +1,6 @@
 import pytest
-
 from pandas.testing import assert_frame_equal
+
 from zospy.analyses.psf import huygens_psf
 
 
