### 2023-12-12 3afdba8
Fix reversed row index in datagrids (#56)
---------

Co-authored-by: LucVV <l.van_vught@lumc.nl>
Co-authored-by: jwmbeenakker <j.w.m.beenakker@lumc.nl>
