### 2023-12-13 5255bc4
Apply suggestions from code review relating to assert_frame_equal
Co-authored-by: Luc van Vught <31657810+LucVV@users.noreply.github.com>
