+++ tests/analyses/test_mtf.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pytest
+from pandas.testing import assert_frame_equal
 
 from zospy.analyses.mtf import (
     fft_through_focus_mtf,
@@ -119,4 +120,4 @@ class TestHuygensMTF:
             maximumfrequency=maximumfrequency,
         )
 
-        assert np.allclose(result.Data.astype(float), expected_data.Data.astype(float), rtol=1e-3)
+       assert_frame_equal(result.Data.astype(float), expected_data.Data.astype(float))
