### 2023-12-15 e400943
Changing naming of method variables.
