### 2023-12-15 7fb3f88
Remove bug fix for mtftype, unnecessary for huygens_mtf.
