### 2023-12-15 ee11f8c
Adding requested variable name changes to unit tests and test reference data specification.
