+++ scripts/generate_test_reference_data/tests.yaml
@@ -36,11 +36,11 @@ tests:
     test: test_huygens_mtf_returns_correct_result
     parametrized: [pupilsampling, imagesampling, imagedelta, mtftype, maximumfrequency]
     parameters:
-      - pupilsampling: 64x64
-        imagesampling: 64x64
-        imagedelta: 0.0
+      - pupil_sampling: 64x64
+        image_sampling: 64x64
+        image_delta: 0.0
         mtftype: Modulation
-        maximumfrequency: 150.0
+        maximum_frequency: 150.0
   - model: polarized_system
     analysis: polarization.polarization_pupil_map
     file: test_polarization.py
