### 2023-12-15 30036dc
Forgot to change variable names in test specification
