### 2023-12-15 ef9f57c
Clarifying documentation of maximum_frequency in huygens_mtf method.
