+++ zospy/analyses/mtf.py
@@ -275,7 +275,9 @@ def huygens_mtf(
     mtftype : zospy.constants.Analysis.Settings.Mtf.MtfTypes.Modulation
         The MTF type (e.g. `Modulation`) that is calculated.
     maximum_frequency : float
-        The maximum frequency at which the MTF is calculated in lp/mm. Defaults to 150.0.
+        The maximum frequency at which the MTF is calculated.
+        Units are either cycles/mm or cycles/mr, depending on system setting).
+        Defaults to 150.0, which is more appropriate when units are set to cycles/mm.
     use_polarization : bool
         Use polarization. Defaults to False.
     use_dashes : bool
