+++ CHANGELOG.md
@@ -15,6 +15,8 @@ ZOS-API can also be added in patch releases.
 
 ### Changed
 
+- Changed license to MIT (#57) - 2023-12-22
+
 ### Deprecated
 
 ### Removed
