### 2023-12-22 b865168
Mention PR #58 in the changelog
