+++ CHANGELOG.md
@@ -15,7 +15,7 @@ ZOS-API can also be added in patch releases.
 
 ### Changed
 
-- Changed license to MIT (#57) - 2023-12-22
+- Changed license to MIT (#57, #58) - 2023-12-22
 
 ### Deprecated
 
