### 2023-12-22 94aebd9
Update license to MIT (#58)
* Update license to MIT

* Update CHANGELOG.md

---------

Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>
