### 2023-12-22 18caf2d
Fix raysandspots tests on systems with a dot as decimal separator (#59)
* Use system dependent decimal point in result keys
