+++ .github/workflows/ci.yml
@@ -80,7 +80,7 @@ jobs:
         - uses: actions/checkout@v3
         - name: Check for file updates
           id: changed-files
-          uses: tj-actions/changed-files@v37
+          uses: tj-actions/changed-files@v41
           with:
             files: CHANGELOG.md
         - name: Fail if nothing changed
