+++ zospy/api/__init__.py
@@ -1,0 +1,8 @@
+"""ZOS-API connection management.
+
+This module contains functions for initialization of and communication with the ZOS-API.
+"""
+
+# from zospy.api import codecs
+#
+# __all__ = ("codecs",)
