### 2024-01-18 f2651e5
Update dependency and compatibility information in README
