### 2024-01-18 6b88f4e
Restore old save behavior where a warning is issued if _OpenFile is None
