### 2024-01-18 22a8f94
Add notes about fixed issues to changelog
