+++ zospy/api/__init__.py
@@ -2,7 +2,3 @@
 
 This module contains functions for initialization of and communication with the ZOS-API.
 """
-
-# from zospy.api import codecs
-#
-# __all__ = ("codecs",)
