+++ tests/analyses/test_raysandspots.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 
 from zospy.analyses.raysandspots import ray_fan, single_ray_trace
+from zospy.api.config import DECIMAL_POINT
 
 
 class TestSingleRayTrace:
@@ -111,12 +112,12 @@ class TestRayFan:
         )
 
         assert np.allclose(
-            result.Data["Tangential fan, field number 1 = 0,0000 (deg)"],
-            expected_data.Data["Tangential fan, field number 1 = 0,0000 (deg)"],
+            result.Data[f"Tangential fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
+            expected_data.Data[f"Tangential fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
         )
         assert np.allclose(
-            result.Data["Sagittal fan, field number 1 = 0,0000 (deg)"],
-            expected_data.Data["Sagittal fan, " "field number 1 = 0," "0000 (deg)"],
+            result.Data[f"Sagittal fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
+            expected_data.Data[f"Sagittal fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
         )
 
     @pytest.mark.parametrize(
@@ -139,10 +140,10 @@ class TestRayFan:
         )
 
         assert np.allclose(
-            result.Data["Tangential fan, field number 1 = 0,0000 (deg)"],
-            reference_data.Data["Tangential fan, field number 1 = 0,0000 (deg)"],
+            result.Data[f"Tangential fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
+            reference_data.Data[f"Tangential fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
         )
         assert np.allclose(
-            result.Data["Sagittal fan, field number 1 = 0,0000 (deg)"],
-            reference_data.Data["Sagittal fan, field number 1 = 0,0000 (deg)"],
+            result.Data[f"Sagittal fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
+            reference_data.Data[f"Sagittal fan, field number 1 = 0{DECIMAL_POINT}0000 (deg)"],
         )
