### 2024-01-19 2944d96
Small improvements to changelog and readme
