+++ .github/workflows/python-test-publish.yml
@@ -23,9 +23,9 @@ jobs:
       contents: read
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v5
       with:
         python-version: '3.x'
     - name: Install dependencies
@@ -35,7 +35,7 @@ jobs:
     - name: Build package
       run: python -m build
     - name: Upload dist
-      uses: actions/upload-artifact@v3.1.2
+      uses: actions/upload-artifact@v4
       with:
         name: zospy_dist
         path: "dist/"
@@ -47,12 +47,12 @@ jobs:
 
     steps:
     - name: Download dist
-      uses: actions/download-artifact@v2.1.1
+      uses: actions/download-artifact@v4
       with:
         name: zospy_dist
         path: dist
     - name: Publish package
-      uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
+      uses: pypa/gh-action-pypi-publish@v1.8.11
       with:
         user: __token__
         password: ${{ secrets.TESTPYPI_API_TOKEN }}
@@ -70,13 +70,19 @@ jobs:
 
     steps:
     - name: Download dist
-      uses: actions/download-artifact@v2.1.1
+      uses: actions/download-artifact@v4
       with:
         name: zospy_dist
         path: dist
     - name: Upload wheels to GitHub release
-      uses: svenstaro/upload-release-action@2.5.0
+      uses: svenstaro/upload-release-action@2.7.0
       with:
         tag: ${{ github.ref }}
         file: "dist/*.whl"
+        file_glob: true
+    - name: Upload dist to GitHub release
+      uses: svenstaro/upload-release-action@2.7.0
+      with:
+        tag: ${{ github.ref }}
+        file: "dist/*.tar.gz"
         file_glob: true
