### 2024-01-19 b2e2b88
Update compatibility information and connection FAQ
