+++ docs/compatibility_table.md
@@ -1,9 +1,9 @@
-| OpticStudio | Python 3.9  | Python 3.10 | Python 3.11 |
-|-------------|:-----------:|:-----------:|:-----------:|
-| 24.1.0      | [✔][24.1.0] | [✔][24.1.0] | [✔][24.1.0] |
-| 23.2.1      | [✔][23.2.1] | [✔][23.2.1] | [✔][23.2.1] |
-| 23.1.0      | [✔][23.1.0] | [✔][23.1.0] | [✔][23.1.0] |
-| 20.3.2      | [ℹ][20.3.2] | [ℹ][20.3.2] | [ℹ][20.3.2] |
+| OpticStudio | Python 3.9  | Python 3.10 | Python 3.11 | Python 3.12 |
+|-------------|:-----------:|:-----------:|:-----------:|:-----------:|
+| 24.1.0      | [✔][24.1.0] | [✔][24.1.0] | [✔][24.1.0] | [✔][24.1.0] |
+| 23.2.1      | [✔][23.2.1] | [✔][23.2.1] | [✔][23.2.1] |             |
+| 23.1.0      | [✔][23.1.0] | [✔][23.1.0] | [✔][23.1.0] |             |
+| 20.3.2      | [ℹ][20.3.2] | [ℹ][20.3.2] | [ℹ][20.3.2] | [ℹ][20.3.2] |
 
 :::{dropdown} Table legend
 ✔: This version works without problems.  
