### 2024-01-19 b4e8734
Configure trusted publishers for PyPI and TestPyPI
