+++ .github/workflows/python-publish.yml
@@ -56,7 +56,6 @@ jobs:
       uses: pypa/gh-action-pypi-publish@v1.8.11
       with:
         user: __token__
-        password: ${{ secrets.PYPI_API_TOKEN }}
         
   upload_to_github:
     name: Upload to GitHub
