### 2024-01-19 b6f4961
XFail test_get_system in extension mode
Due to (presumably) a bug in the ZOS-API, the primary system is not available at index 0.
