### 2024-01-19 ac06bfd
Add permissions required for trusted publishing
