### 2024-01-19 231d88c
Change note about Python.NET 3 and add some badges
