### 2024-01-19 b78a44c
Changes for 1.2.0 from LUMC GitLab (#62)
* Add Python.NET codecs
* Add unified `ZOS.connect` method
* Add `ZOS.disconnect` method to disconnect from OpticStudio
* Add support for OpticStudio 2024 R1 and Python 3.12
* Allow relative paths and check for their existence in `OpticStudioSystem.load` and `OpticStudioSystem.save_as`

---------

Co-authored-by: LucVV <l.van_vught@lumc.nl>
Co-authored-by: jwmbeenakker <j.w.m.beenakker@lumc.nl>
Co-authored-by: Jan-Willem Beenakker <90187768+jwmbeenakker@users.noreply.github.com>
