+++ zospy/zpcore.py
@@ -4,11 +4,14 @@ import locale
 import logging
 import warnings
 import weakref
+from os import PathLike
+from typing import Literal
 
 from semver.version import Version
 
 from zospy.api import _ZOSAPI, constants
 from zospy.api.apisupport import load_zosapi, load_zosapi_nethelper
+from zospy.utils.pyutils import abspath
 
 logger = logging.getLogger(__name__)
 
@@ -190,16 +193,18 @@ class OpticStudioSystem:
         """Set the optical system to non-sequential mode if it is not already."""
         return self._System.MakeNonSequential()
 
-    def load(self, filepath: str, saveifneeded: bool = False):
+    def load(self, filepath: str | PathLike, saveifneeded: bool = False):
         """Load an earlier defined Zemax file.
 
         Parameters
         ----------
-        filepath: str
-            The filepath that should be loaded
+        filepath: str | PathLike
+            Path to the file that should be loaded
         saveifneeded: bool
             Defines if the current file is saved before opening the new file. Defaults to False.
         """
+        filepath = abspath(filepath)
+
         logger.debug("Opening {} with SaveIfNeeded set to {}".format(filepath, saveifneeded))
 
         self._System.LoadFile(filepath, saveifneeded)
@@ -222,14 +227,16 @@ class OpticStudioSystem:
 
         logger.info("Opened new file")
 
-    def save_as(self, filepath: str):
+    def save_as(self, filepath: str | PathLike):
         """Save the current session under a specified name.
 
         Parameters
         ----------
-        filepath: str
+        filepath: str | PathLike
             The filepath where the system should be saved. Note that a partial path or relative path does not work.
         """
+        filepath = abspath(filepath, check_directory_only=True)
+
         logger.debug("Saving open session as {}".format(filepath))
 
         self._System.SaveAs(filepath)
@@ -240,8 +247,8 @@ class OpticStudioSystem:
     def save(self) -> bool:
         """Save the current OpticStudio session.
 
-        If the file name for the current session is not known (e.g. when a new file was created), a warning is raised
-        and the file is not saved. On these occurences, save_as() should be used once.
+        If the file name for the current session is not known (e.g. when a new file was created), a warning is raised.
+        On these occurences, save_as() should be used once.
 
         Returns (bool): A boolean indicating if the saving was successful.
 
@@ -250,16 +257,14 @@ class OpticStudioSystem:
 
         if self._OpenFile is None:
             warnings.warn(
-                "No file name has been defined for the current system, the current session has not been "
-                "saved. Please use the save_as() function before using save."
+                "The optical system has not been saved because no file name has been defined for the current system."
+                "Please use the 'save_as' function before using 'save'."
             )
-            logger.critical("Could not save file")
+            logger.critical("Could not save file because the path has not been set using save_as")
 
-            return False
-        else:
-            self._System.Save()
-            logger.info("File saved")
-            return True
+        self._System.Save()
+        logger.info("File saved")
+        return True
 
     def close(self, saveifneeded: bool = False) -> bool:
         """Close the optical system. Only works on secondary systems (See OpticStudio documentation).
@@ -276,7 +281,7 @@ class OpticStudioSystem:
 
         Returns
         -------
-        zospy.core.OpticStudioSystem
+        zospy.zpcore.OpticStudioSystem
             A ZOSPy OpticStudioSystem instance. Should be sequential.
[…]
