+++ zospy/__init__.py
@@ -1,6 +1,6 @@
 """A Python package to communicate with Zemax OpticStudio through its API."""
 
-__version__ = "1.1.2"
+__version__ = "1.2.0"
 
 import logging
 

[…]
