### 2024-01-23 #5756 @plaplant
@LucVV thanks for your patience, and apologies for not responding sooner. I have reached out via email to @alvesjnr, and have not heard back. I think at this point, unfortunately the next course of action is to find another suitable reviewer, as JOSS policy is to have two reviewers full approve the submission before acceptance. I apologize for the additional delay this will cause, but I will begin reaching out to potential reviewers right away so we can move this forward.
