### 2024-01-26 #5756 @LucVV
I have slightly updated figure 1 to make sure it exactly matches the images in the [example on the documentation website](https://zospy.readthedocs.io/en/latest/examples/Simple%20thick%20lens/Simple%20thick%20lens.html). I hope that is not an issue. The wording or message of the paper did not change in any way.
