### 2024-02-05 02863af
Add conda badges and command to README
