### 2024-02-06 9cfdc9d
Remove conda platform badge
