### 2024-02-28 #5756 @plaplant
@LucVV I just wanted to put a small (non) update here. I am sorry to say that I am still looking for an alternative reviewer. If by chance you know someone who might be qualified to review and does not have a conflict of interest (defined for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)), then please suggest them and I would be happy to have them review. Thanks again for your patience!
