### 2024-02-28 #5756 @jwmbeenakker
Thank you for the update @plaplant . We think omnistic could be a potential reviewer. He has contributed in the past to ZOSPy, but if I look at the COI policy, this should not be a problem.
