### 2024-03-01 c130f97
Update config.py
fixed/ worked around locale.setlocale-problem
