### 2024-03-01 04f45d9
added mtfclasses with FFT_MTF_MAP and FFT_MTF
added mtfclasses with FFT_MTF_MAP and FFT_MTF classes featuring some helpers for plotting with matplotlib.pyplot
