### 2024-03-01 228ad39
enhancements to LDE
added .SurfaceByComment and .SurfaceByRegex
