### 2024-03-07 4e0d6c8
Use locale.getencoding instead of locale.getpreferredencoding
locale.getpreferredencoding returns utf-8 if Python is put in UTF-8 mode. locale.getencoding always returns the system default encoding.
