### 2024-03-07 0bb5d9e
Fix failing unit tests for german locale
