+++ Paper/paper.md
@@ -1,108 +1,0 @@
----
-title: 'ZOSPy: optical ray tracing in Python through OpticStudio'
-tags:
-  - Python
-  - Ray tracing
-  - OpticStudio
-  - Zemax
-authors:
-  - name: Luc van Vught
-    affiliation: "1, 2"
-    corresponding: true
-    equal-contrib: true
-    orcid: 0000-0001-8290-9071
-  - name: Corné Haasjes
-    affiliation: "1, 2, 3"
-    equal-contrib: true
-    orcid: 0000-0003-0187-4116
-  - name: Jan-Willem M. Beenakker
-    affiliation: "1, 2, 3"
-    orcid: 0000-0003-0479-5587
-affiliations:
-  - name: Department of Ophthalmology, Leiden University Medical Center, Leiden, the Netherlands
-    index: 1
-  - name: Department of Radiology, C.J. Gorter MRI Center, Leiden University Medical Center, Leiden, the Netherlands
-    index: 2
-  - name: Department of Radiation Oncology, Leiden University Medical Center, Leiden, the Netherlands
-    index: 3
-date: 07 June 2023
-bibliography: paper.bib
----
-
-# Summary
-Zemax OpticStudio (Ansys, Inc) is a commonly used software package for designing optical setups and performing ray tracing simulations. It offers an Application Programming Interface (API) but interacting with this API is complex. Consequently, current ray tracing simulations generally require substantial manual user interaction, which in turn hampers the sharing of methods between scientists. We have therefore developed `ZOSPy`, a Python package that provides an accessible interface as well as unit tests. As a result, `ZOSPy` enables scientists to focus more on optical modelling instead of coding and contributes to open science as optical setups and analyses can easily be shared amongst users. 
-
-# Statement of need
-Ray tracing simulations are widely used to design, optimize and analyze optical systems. Its applications are diverse, ranging from designing spectrometers [@Naeem2022] or telescopes [@Zhang2023], to understanding the optics of the human eye [@Simpson2020; @vanVught2022]. Moreover, in ophthalmology, ray tracing is used to optimize the outcomes of cataract surgery [@Canovas2011; @Artal2023] and evaluate the accuracy of ocular radiotherapy [@Jaarsma2023]. These optical simulations are often performed in OpticStudio, which offers a powerful set of tools to design, optimize and evaluate optical systems.
-
-Although OpticStudio offers an API, the `ZOS-API`, using this API in Python is complex and time-consuming. It involves, for example, establishing a connection with the API through the .NET framework, casting between .NET and Python datatypes, identifying which constants need to be set in specific cases, and working around  non-uniform methods of parsing the output [@GettingStartedWithZOSAPI]. This leads to studies which, in practice, largely rely on user interaction and therefore generally only assess a small number of optical systems. It furthermore hinders open science initiatives as sharing of methods is limited to screenshots [@Hong2011], simple tables [@Naeem2022] or optical systems saved in the proprietary file format of OpticStudio [@Polans2018]. 
-
-With `ZOSPy`, we aim to provide an accessible interface to the OpticStudio API, enabling the user to focus on optical modelling instead of complex coding. Furthermore, we hope that the easy integration of the ray tracing workflow from optical design to analyses into Python scripts or Jupyter Notebooks facilitates open science. 
-
-# Functionality
[…]
