+++ Paper/paper.bib
@@ -1,117 +1,0 @@
-
-@article{Artal2023,
-  title = {Inverted meniscus intraocular lens as a better optical surrogate of the crystalline lens},
-  author = {Artal, Pablo and Ginis, Harilaos and Christaras, Dimitrios and Villegas, Eloy A and Tabernero, Juan and Prieto, Pedro M},
-  journal = {Biomedical Optics Express},
-  volume = {14},
-  number = {5},
-  pages = {2129--2137},
-  year = {2023},
-  publisher = {Optica Publishing Group},
-  doi = {10.1364/BOE.490089}
-}
-
-@article{Canovas2011,
-  title = {Customized eye models for determining optimized intraocular lenses power},
-  author = {Canovas, Carmen and Artal, Pablo},
-  journal = {Biomedical optics express},
-  volume = {2},
-  number = {6},
-  pages = {1649--1662},
-  year = {2011},
-  publisher = {Optica Publishing Group},
-  doi = {10.1364/BOE.2.001649}
-}
-
-@article{Hong2011,
-  title = {Ray-tracing optical modeling of negative dysphotopsia},
-  author = {Hong, Xin and Liu, Yueai and Karakelle, Mutlu and Masket, Samuel and Fram, Nicole R},
-  journal = {Journal of biomedical optics},
-  volume = {16},
-  number = {12},
-  pages = {125001--125001},
-  year = {2011},
-  publisher = {Society of Photo-Optical Instrumentation Engineers},
-  doi = {10.1117/1.3656745}
-}
-
-@article{Jaarsma2023,
-  title = {Comparison of Magnetic Resonance Imaging--Based and Conventional Measurements for Proton Beam Therapy of Uveal Melanoma},
-  author = {Jaarsma-Coes, Myriam G and Ferreira, Teresa A and Marinkovic, Marina and Vu, TH Khanh and van Vught, Luc and van Haren, Guido R and Rodrigues, Myra F and Klaver, Yvonne LB and Verbist, Berit M and Luyten, Gregorius PM and others},
-  journal = {Ophthalmology Retina},
-  volume = {7},
-  number = {2},
-  pages = {178--188},
-  year = {2023},
-  publisher = {Elsevier},
-  doi = {10.1016/j.oret.2022.06.019}
-}
-
-@Article{Naeem2022,
-  title = {Design Simulation and Data Analysis of an Optical Spectrometer},
-  author = {Naeem, Muddasir and Imran, Tayyab and Hussain, Mukhtar and Bhatti, Arshad Saleem},
-  journal = {Optics},
-  volume = {3},
-  year = {2022},
-  number = {3},
-  pages = {304--312},
-  url = {https://www.mdpi.com/2673-3269/3/3/28},
-  issn = {2673-3269},
-  doi = {10.3390/opt3030028}
-}
-
-@article{Polans2018,
-  author = {James Polans and Bart Jaeken and Ryan P. McNabb and Pablo Artal and Joseph A. Izatt},
-  journal = {Optica},
-  keywords = {Crystalline lens; Eye models; Gradient index optics; Retina},
-  number = {11},
-  pages = {1461--1461},
-  publisher = {Optica Publishing Group},
-  title = {Asymmetric wide-field optical model of the human eye with tilted and decentered crystalline lens that reproduces experimentally measured aberrations: errata},
-  volume = {5},
-  month = {Nov},
-  year = {2018},
-  url = {https://opg.optica.org/optica/abstract.cfm?URI=optica-5-11-1461},
-  doi = {10.1364/OPTICA.5.001461}
-}
-
-@article{Simpson2020,
[…]
