### 2024-03-07 2e44c8e
Fix unsupported locale exception on import (#69)
* Use setlocale to determine the currently set locale

The same method is used in the Python locale module as well: https://github.com/python/cpython/blob/5c69f60ae1859e53f858e65636d1c6d83873a963/Lib/locale.py#L682C9-L690C41

* Use locale.getencoding instead of locale.getpreferredencoding

locale.getpreferredencoding returns utf-8 if Python is put in UTF-8 mode. locale.getencoding always returns the system default encoding.

* Fix test_get_txtfile_encoding_returns_correct_result

* Fix failing unit tests for german locale
