### 2024-03-07 4d1eb76
Pull changes from LUMC GitLab (custom `dir` for `Analysis` objects) (#68)
* synchronize GitLab with GitHub

Co-authored-by: LucVV <l.van_vught@lumc.nl>
Co-authored-by: jwmbeenakker <j.w.m.beenakker@lumc.nl>
Co-authored-by: crnh <30109443+crnh@users.noreply.github.com>

---------

Co-authored-by: LucVV <l.van_vught@lumc.nl>
Co-authored-by: jwmbeenakker <j.w.m.beenakker@lumc.nl>
