### 2024-03-11 3301d40
Use locale.getpreferredencoding when Python < 3.11
