### 2024-03-11 af163a8
Also use getpreferredencoding in unit tests for Python < 3.11
