### 2024-03-11 fcd6817
Use localized decimal point and string conversions
