+++ zospy/__init__.py
@@ -1,6 +1,6 @@
 """A Python package to communicate with Zemax OpticStudio through its API."""
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 
 import logging
 
