+++ .github/workflows/python-test-publish.yml
@@ -59,9 +59,9 @@ jobs:
       uses: pypa/gh-action-pypi-publish@v1.8.11
       with:
         user: __token__
-        repository_url: https://test.pypi.org/legacy/
+        repository-url: https://test.pypi.org/legacy/
         # If the release is already present, overwrite it
-        skip_existing: true
+        skip-existing: true
         
   publish_to_github:
     runs-on: ubuntu-latest
