### 2024-03-11 b0afa57
Version 1.2.1 (#71)
* Prepare for version 1.2.1

* Update test publish action
