### 2024-03-15 #5756 @plaplant
@Omnistic thanks for agreeing to review! When you get a chance, please reply to this thread with the comment:

As an editorial note, @Omnistic has contributed to this repo in the past. However, these contributions seem relatively minor, he is not a co-author of the manuscript, and he complies with all of the components of the stated [JOSS Conflict of Interest Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). As such, I believe he can serve as a reviewer for this submission.
