### 2024-03-20 #5756 @Omnistic
I am generally happy with the paper. I only have one small remark about the caption of Figure 1. When the authors write: _"the image plane needs a 1.9 mm shift towards the lens to be in focus"_, they clearly know what they are talking about. However, this is probably not the case of every reader, myself included. Can you please give a short explanation?
