### 2024-03-21 f5637ea
Update example code in paper
