### 2024-03-21 714c45a
updated figure caption
with a more clear description
