### 2024-03-22 e1d208e
Add non-breaking space between number and unit in caption of Figure 1
