### 2024-03-22 #5756 @jwmbeenakker
@Omnistic Thank you for your review and spotting the 1.9mm mistake (it was based on an earlier version of the manuscript, before the review of brandondube).
We slightly modified the example, figure and associated caption making it more straightforward for the reader to understand.

@plaplant Is there anything else we need to do?
