### 2024-03-24 #5756 @plaplant
@Omnistic @brandondube thanks very much for your reviews of the software package! We appreciate your feedback on the software repository and suggestions for improvement.

@LucVV now that both reviewers have recommended acceptance of the submission, we can move forward with the final acceptance checks. Please follow these steps:
- If the software has changed in the course of the review (besides changes to the paper), please make a new tagged release of the repository.
- Once that is done, please archive is (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or some other long-term hosting platform).
  - When performing this step, please make sure the authors of the archive match the authors of the paper (including ORCIDs).
- Please post the version number and DOI of the archive in this thread.

Let me know if you have any questions!

@jwmbeenakker I see that you tried to set the version for the software as v1.2.1. Is this the latest official release, that includes all changes that took place as part of the review? If so, I will set that as the official version. If not, please make a new release as described above and let me know the new version number. Thanks!
