### 2024-03-25 #5756 @jwmbeenakker
@plaplant great news :-)
v1.2.1 is indeed the latest official release which includes all changes that were part of the review.
We also archived it on zenodo: https://doi.org/10.5281/zenodo.10805414
