### 2024-04-03 #5756 @plaplant
@jwmbeenakker it looks like the paper bibliography includes two publications that are not explicitly referenced in the paper:
- Hong et al. (2011) (DOI: `10.1117/1.3656745`)
- Polans et al. (2018) (DOI: `10.1364/OPTICA.5.001461`)

These entries should either be removed from the `paper.bib` file, or cited in the paper. Once this has been resolved, I think the submission is ready for acceptance. Thanks very much for your patience!
