### 2024-04-03 1f3b37e
Update paper.md
readded Hong2011 reference
