+++ paper/paper.bib
@@ -96,21 +96,6 @@
   doi = {10.3390/opt3030028}
 }
 
-@article{Polans2018,
-  author = {James Polans and Bart Jaeken and Ryan P. McNabb and Pablo Artal and Joseph A. Izatt},
-  journal = {Optica},
-  keywords = {Crystalline lens; Eye models; Gradient index optics; Retina},
-  number = {11},
-  pages = {1461--1461},
-  publisher = {Optica Publishing Group},
-  title = {Asymmetric wide-field optical model of the human eye with tilted and decentered crystalline lens that reproduces experimentally measured aberrations: errata},
-  volume = {5},
-  month = {Nov},
-  year = {2018},
-  url = {https://opg.optica.org/optica/abstract.cfm?URI=optica-5-11-1461},
-  doi = {10.1364/OPTICA.5.001461}
-}
-
 @article{Simpson2020,
   title = {Intraocular lens far peripheral vision: image detail and negative dysphotopsia},
   author = {Simpson, Michael J},
