### 2024-04-04 bc81f84
Update citation file with JOSS paper
